+++ Sources/RIDIFoundation/Extensions/Foundation/UserDefaults+Codable.swift
@@ -1,17 +1,15 @@
 import Foundation
 
 extension UserDefaults {
-    struct _DecodableJSONRoot<T: Decodable>: Decodable {
-         let root: T
-     }
+struct _DecodableJSONRoot<T: Decodable>: Decodable {
+        let root: T
+    }
 
-     struct _EncodableJSONRoot<T: Encodable>: Encodable {
-         let root: T
-     }
+    struct _EncodableJSONRoot<T: Encodable>: Encodable {
+        let root: T
+    }
 
-    open func ridi_object<T>(forKey defaultName: String) throws -> T? where T: Decodable {
-        let value: Any? = object(forKey: defaultName)
-
+    static func _decode<T: Decodable>(_ value: Any?) throws -> T? {
         switch value {
         case let value as NSNumber:
             return value as? T
@@ -41,20 +39,29 @@ extension UserDefaults {
         }
     }
 
-    open func ridi_set<T>(_ value: T?, forKey defaultName: String) throws where T: Encodable {
+    static func _encode<T: Encodable>(_ value: T?) throws -> Any? {
         switch value {
         case let value as NSNumber:
-            set(value, forKey: defaultName)
+            return value
         case let value as NSString:
-            set(value, forKey: defaultName)
+            return value
         case let value as NSDate:
-            set(value, forKey: defaultName)
+            return value
         default:
             let archiver = NSKeyedArchiver(requiringSecureCoding: true)
             try archiver.encodeEncodable(value, forKey: NSKeyedArchiveRootObjectKey)
-            let data = archiver.encodedData
 
-            set(data, forKey: defaultName)
+            return archiver.encodedData
         }
+    }
+
+    open func ridi_object<T>(forKey defaultName: String) throws -> T? where T: Decodable {
+        let value: Any? = object(forKey: defaultName)
+
+        return try Self._decode(value)
+    }
+
+    open func ridi_set<T>(_ value: T?, forKey defaultName: String) throws where T: Encodable {
+        return try set(Self._encode(value), forKey: defaultName)
     }
 }
